// src/net/bindings/types.ts
export const BeeState = __t.object("BeeState", {
  playerId: __t.u32(),
  cell: __t.u32(),
  x: __t.f32(),
  y: __t.f32(),
  z: __t.f32(),
  yaw: __t.i16(),
  pitch: __t.i16(),
  aimYaw: __t.i16(),
  aimPitch: __t.i16(),
  tick: __t.u32(),
  hp: __t.u16(),
  maxHp: __t.u16(),
  flags: __t.u8(),
});
export type BeeState = __Infer<typeof BeeState>;

export const CombatEvent = __t.object("CombatEvent", {
  cell: __t.u32(),
  kind: __t.u8(),
  sourceKind: __t.u8(),
  sourceId: __t.u32(),
  targetKind: __t.u8(),
  targetId: __t.u32(),
  value: __t.u16(),
  aux: __t.u16(),
});
export type CombatEvent = __Infer<typeof CombatEvent>;
// ...
export const FlowerPatch = __t.object("FlowerPatch", {
  patchId: __t.u32(),
  cell: __t.u32(),
  pollen: __t.u16(),
  storedAt: __t.f64(),
});
export type FlowerPatch = __Infer<typeof FlowerPatch>;

export const FlyBrain = __t.object("FlyBrain", {
  flyId: __t.u32(),
  nestId: __t.u8(),
  active: __t.bool(),
  goalX: __t.f32(),
  goalY: __t.f32(),
  goalZ: __t.f32(),
  orbitSign: __t.i8(),
  nextThinkAtMs: __t.u64(),
  nextSpitAtMs: __t.u64(),
  nextSummonAtMs: __t.u64(),
  lastHitBy: __t.u32(),
});
export type FlyBrain = __Infer<typeof FlyBrain>;

export const FlyState = __t.object("FlyState", {
  flyId: __t.u32(),
  cell: __t.u32(),
  x: __t.f32(),
  y: __t.f32(),
  z: __t.f32(),
  vx: __t.i16(),
  vy: __t.i16(),
  vz: __t.i16(),
  yaw: __t.i16(),
  pitch: __t.i16(),
  kind: __t.u8(),
  hp: __t.u16(),
  state: __t.u8(),
  target: __t.u32(),
  tick: __t.u32(),
});
export type FlyState = __Infer<typeof FlyState>;

export const HiveState = __t.object("HiveState", {
  hiveId: __t.u8(),
  honeyDelivered: __t.u32(),
  visits: __t.u32(),
});
export type HiveState = __Infer<typeof HiveState>;

export const ModuleActivation = __t.object("ModuleActivation", {
  id: __t.u64(),
  playerId: __t.u32(),
  slot: __t.u8(),
  targetKind: __t.u8(),
  targetId: __t.u32(),
  nextCycleAtMs: __t.u64(),
  stopAfterCycle: __t.bool(),
});
export type ModuleActivation = __Infer<typeof ModuleActivation>;

export const MyModuleRow = __t.object("MyModuleRow", {
  id: __t.u64(),
  slot: __t.u8(),
  targetKind: __t.u8(),
  targetId: __t.u32(),
  nextCycleAtMs: __t.u64(),
  stopAfterCycle: __t.bool(),
});
export type MyModuleRow = __Infer<typeof MyModuleRow>;

export const MyModules = __t.object("MyModules", {});
export type MyModules = __Infer<typeof MyModules>;

export const MyQuestRow = __t.object("MyQuestRow", {
  id: __t.u64(),
  questId: __t.u16(),
  progress: __t.u32(),
  completions: __t.u16(),
});
export type MyQuestRow = __Infer<typeof MyQuestRow>;

export const MyQuests = __t.object("MyQuests", {});
export type MyQuests = __Infer<typeof MyQuests>;

export const MyStats = __t.object("MyStats", {});
export type MyStats = __Infer<typeof MyStats>;

export const MyStatsRow = __t.object("MyStatsRow", {
  playerId: __t.u32(),
  honey: __t.u32(),
  cargo: __t.u16(),
  cargoGold: __t.u16(),
  energy: __t.f32(),
  energyAt: __t.f64(),
  hp: __t.u16(),
  lensLevel: __t.u8(),
  gatlingLevel: __t.u8(),
  quiverLevel: __t.u8(),
  brushLevel: __t.u8(),
  cargoLevel: __t.u8(),
  wingsLevel: __t.u8(),
  armorLevel: __t.u8(),
  antennaLevel: __t.u8(),
  nectarLevel: __t.u8(),
  dockedHive: __t.u8(),
  hivesVisited: __t.u16(),
  achievements: __t.u32(),
  deaths: __t.u32(),
  savedX: __t.f32(),
  savedY: __t.f32(),
  savedZ: __t.f32(),
  savedFlags: __t.u8(),
});
export type MyStatsRow = __Infer<typeof MyStatsRow>;
// ...
export const PlayerProfile = __t.object("PlayerProfile", {
  playerId: __t.u32(),
  identity: __t.identity(),
  name: __t.string(),
  homeHive: __t.u8(),
  isDemo: __t.bool(),
});
export type PlayerProfile = __Infer<typeof PlayerProfile>;

export const PlayerScore = __t.object("PlayerScore", {
  playerId: __t.u32(),
  honeyTotal: __t.u32(),
  kills: __t.u32(),
  pollenTotal: __t.u32(),
});
export type PlayerScore = __Infer<typeof PlayerScore>;
// ...
export const WorldWeather = __t.object("WorldWeather", {
  id: __t.u8(),
  weather: __t.string(),
  sinceWorldSeconds: __t.f64(),
});
export type WorldWeather = __Infer<typeof WorldWeather>;